refactor(logger): refactor methods to better naming
